Issue 235: Use SharedTheme for JAXB unmarshalling
